Clean up CVXPY integration: remove old code and rename working versions
- Renamed fixed_* functions to main versions (json_to_lean_code, generate_examples)
- Renamed Fixed* classes to clean names (SExprToLeanTranslator, JSONToLeanConverter)
- Renamed Fixed*.lean files to clean names (Quadratic.lean, SimpleLP.lean, Portfolio.lean)
- Updated all references and imports to use clean names
- Removed "Fixed" prefixes from docstrings and print statements

The integration now has a clean, production-ready API.
