+++ CvxLean/Examples/CVXPY/Quadratic.lean
