+++ CvxLean/Examples/CVXPY/SimpleLP.lean
